app/requirements.txt
asgiref==3.8.1
click==8.1.7
Django==5.1
django-prometheus==2.3.1
djangorestframework==3.15.2
h11==0.14.0
prometheus_client==0.20.0
sqlparse==0.5.1
uvicorn==0.30.6
